chore: reorganize tests
