+++ tests/README.md
@@ -1,0 +1,20 @@
+# Avrae Tests
+
+The Avrae test suite is comprised of three primary parts: end-to-end (E2E) testing (`e2e/`), unit testing (`unit/`), and
+gamedata simulation testing (`gamedata/`).
+
+## Setup
+
+```shell
+(venv) $ pip install -r tests/requirements.txt
+```
+
+In order to run the gamedata simulation tests,
+
+### Environment Variables
+
+TODO
+
+## Running Tests
+
+TODO
