+++ tests/e2e/bot_test.py
