+++ tests/e2e/cogs/charGen_test.py
