+++ tests/e2e/cogs/customization_test.py
