+++ tests/e2e/cogs/dice_test.py
