+++ tests/e2e/cogs/lookup_test.py
