+++ tests/e2e/cogs/sheetManager_test.py
