+++ tests/e2e/help_test.py
