+++ tests/unit/coinpurse_test.py
